kept file open
In this version the outfile is kept open. Could further simplify by
changing the a to a w, and then removing the .os section to delete the
file in the top of the program.  The sample file of 4,1000 points runs
in around 8 seconds in non verbose mode on this particular i7.  Would
love to make this faster.